% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width.R
\name{set_equiwidth}
\alias{set_equiwidth}
\title{Set a global equiwidth}
\usage{
set_equiwidth(equiwidth = 1)
}
\arguments{
\item{equiwidth}{A single positive finite numeric value.}
}
\value{
The previous option value, invisibly.
}
\description{
Set a global default for the \code{equiwidth} argument used by \code{\link[=get_width]{get_width()}}.
}
