% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise-width.R
\name{set_width}
\alias{set_width}
\title{Set the width standard}
\usage{
set_width(standard = 1)
}
\arguments{
\item{standard}{Numeric. The default width standard. A value of \code{1} (default)
corresponds to a standard width. Increase to make elements thicker,
decrease to make them thinner.}
}
\description{
Sets a global default for the \code{standard} argument in \code{standardise_width()}.
All subsequent calls to \code{standardise_width()} will use this value when
\code{standard = NULL}.
}
