% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise-width.R
\name{set_width_appearance}
\alias{set_width_appearance}
\title{Set a global width appearance appearance}
\usage{
set_width_appearance(x = 1)
}
\arguments{
\item{x}{Numeric. A value of \code{1} (default) corresponds to a appearance width.
Increase to make a wider width appearance, and decrease to make a thinner
width appearance.}
}
\description{
Sets a global default for the \code{appearance} argument in \code{standardise_width()}.
All subsequent calls to \code{standardise_width()} will use this value when
\code{appearance = NULL}.
}
\examples{
set_width_appearance(1)
set_width_appearance(0.75)
set_width_appearance(1.33)
}
\seealso{
\code{\link[=standardise_width]{standardise_width()}}
}
