% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise-width.R
\name{set_width_standard}
\alias{set_width_standard}
\title{Set the width standard}
\usage{
set_width_standard(x)
}
\arguments{
\item{x}{Numeric. A value of \code{1} (default) corresponds to a standard width.
Increase to make elements thicker, decrease to make them thinner.}
}
\description{
Sets a global default for the \code{standard} argument in \code{standardise_width()}.
All subsequent calls to \code{standardise_width()} will use this value when
\code{standard = NULL}.
}
\examples{
set_width_standard(1)
set_width_standard(0.75)
set_width_standard(1.33)
}
\seealso{
\code{\link[=standardise_width]{standardise_width()}}
}
