% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwidth-package.R
\docType{package}
\name{ggwidth-package}
\alias{ggwidth}
\alias{ggwidth-package}
\title{ggwidth: Standardise 'ggplot2' Geom Width}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Standardise the width in 'ggplot2' geoms to appear visually consistent across plots with different numbers of categories, panel dimensions, and orientations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidhodge931/ggwidth}
  \item \url{https://davidhodge931.github.io/ggwidth/}
  \item Report bugs at \url{https://github.com/davidhodge931/ggwidth/issues}
}

}
\author{
\strong{Maintainer}: David Hodge \email{davidhodge931@gmail.com} (\href{https://orcid.org/0000-0002-3868-7501}{ORCID}) [copyright holder]

}
\keyword{internal}
