% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-equiwidth.R
\name{set_equiwidth}
\alias{set_equiwidth}
\title{Set a global width equiwidth}
\usage{
set_equiwidth(equiwidth = 1)
}
\arguments{
\item{equiwidth}{Numeric. Multiplicative factor that controls the width appearance.
A value of \code{1} (default) is the default. Increase to make a wider appearance, and
decrease to make a thinner appearance. If \code{NULL}, uses the value set by \code{set_equiwidth()},
falling back to \code{1}.}
}
\description{
Sets a global default for the \code{equiwidth} argument in \code{get_width()}.
All subsequent calls to \code{get_width()} will use this value when
\code{equiwidth = NULL} falling back to 1.
}
\examples{
set_equiwidth(1)
set_equiwidth(0.75)
set_equiwidth(1.33)
}
\seealso{
\code{\link[=get_width]{get_width()}}
}
