% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise-width.R
\name{set_width}
\alias{set_width}
\title{Set a global width scale}
\usage{
set_width(scale = 1)
}
\arguments{
\item{scale}{Numeric. A value of \code{1} (default) corresponds to a default
width scale. Increase to make a wider width scale, and decrease
to make a thinner width scale.}
}
\description{
Sets a global default for the \code{scale} argument in \code{standardise_width()}.
All subsequent calls to \code{standardise_width()} will use this value when
\code{scale = NULL}.
}
\examples{
set_width(1)
set_width(0.75)
set_width(1.33)
}
\seealso{
\code{\link[=standardise_width]{standardise_width()}}
}
