% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise-width.R
\name{standardise_width}
\alias{standardise_width}
\title{Standardise width to the reference}
\usage{
standardise_width(n = NULL, dodge_n = 1, orientation = "x", ...)
}
\arguments{
\item{n}{Number of categories in the plot.}

\item{dodge_n}{Number of dodge groups.}

\item{orientation}{Orientation ("x" or "y").}

\item{...}{Require named arguments.}
}
\description{
Standardise width to the reference
}
\examples{
library(ggplot2)
library(dplyr)
library(palmerpalmerpenguins)

#' Set global theme and width reference
theme_set(theme_grey() +
            theme(panel.widths = rep(unit(75, "mm"), 2)) +
            theme(panel.heights = rep(unit(50, "mm"), 2)))

#' 1. Standard vertical bar chart
palmerpenguins::penguins |>
  filter(!is.na(sex)) |>
  ggplot(aes(x = sex, fill = species)) +
  geom_bar(
    width = standardise_width(
      n = 2,
      dodge_n = 1,
      orientation = "x"
    )
  )

#' 2. Vertical bar chart with dodging
#' Note: dodge_n = 3 matches the 3 species in the fill aesthetic
palmerpenguins::penguins |>
  filter(!is.na(sex)) |>
  mutate(sex = str_to_sentence(sex)) |>
  ggplot(aes(x = sex, fill = species)) +
  geom_bar(
    position = position_dodge(),
    width = standardise_width(
      n = 2,
      dodge_n = 3,
      orientation = "x"
    )
  )

#' 3. Horizontal bar chart with dodging
#' Uses orientation = "y" to scale against panel height
palmerpenguins::penguins |>
  tidyr::drop_na(sex) |>
  mutate(sex = str_to_sentence(sex)) |>
  ggplot(aes(y = sex, fill = species)) +
  geom_bar(
    position = position_dodge(),
    width = standardise_width(
      n = 2,
      dodge_n = 3,
      orientation = "y",
    )
  )

#' 4. Faceted horizontal bars with free scales
#' Using a reference n ensures bars stay the same thickness across facets
d <- tibble::tibble(
  continent = c("Europe", "Europe", "Europe", "Europe", "Europe",
                "South America", "South America"),
  country = c("AT", "DE", "DK", "ES", "PK", "TW", "BR"),
  value = c(10L, 15L, 20L, 25L, 17L, 13L, 5L)
)

max_n <- d |>
  count(continent) |>
  pull(n) |>
  max()

d |>
  mutate(country = forcats::fct_rev(country)) |>
  ggplot(aes(y = country, x = value)) +
  geom_col(
    width = standardise_width(
      n = max_n,
      dodge_n = 1,
      orientation = "y"
    )
  ) +
  facet_wrap(~continent, scales = "free_y") +
  scale_y_discrete(continuous.limits = c(1, max_n)) +
  coord_cartesian(reverse = "y", clip = "off")

}
