% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width.R
\name{update_equiwidth}
\alias{update_equiwidth}
\title{Update the global equiwidth}
\usage{
update_equiwidth(value = 1)
}
\arguments{
\item{value}{A single positive finite numeric value.}
}
\value{
The previous option value, invisibly.
}
\description{
Update a global default for the \code{value} argument used by \code{\link[=get_width]{get_width()}}.
}
